Split scatter methods into separate files for common, glTF, and USD.
