+++ glue_ar/common/__init__.py
@@ -1,5 +1,6 @@
 from .marching_cubes import add_isosurface_layer_gltf, add_isosurface_layer_usd  # noqa: F401
-from .scatter import add_scatter_layer_gltf, add_scatter_layer_usd  # noqa: F401
+from .scatter_gltf import add_scatter_layer_gltf  # noqa: F401
+from .scatter_usd import add_scatter_layer_usd  # noqa: F401
 from .voxels import add_voxel_layers_gltf, add_voxel_layers_usd  # noqa: F401
 from .scatter_export_options import ARScatterExportOptions  # noqa: F401
 from .volume_export_options import ARIsosurfaceExportOptions, ARVoxelExportOptions  # noqa: F401
