+++ glue_ar/usd_utils.py
@@ -1,17 +1,21 @@
-from typing import Tuple
+from typing import Optional, Tuple
 
-from pxr import Sdf, Usd, UsdShade
+from pxr import Sdf, Usd, UsdGeom, UsdShade
 
-from glue_ar.utils import unique_id
+
+def color_identifier(color: Tuple[int, int, int], opacity: float = 1.0) -> str:
+    return f"{'_'.join(str(c) for c in color)}_{opacity}".replace(".", "_")
 
 
 def material_for_color(stage: Usd.Stage,
                        color: Tuple[int, int, int],
                        opacity: float,
-                       metallic: float = 0.1,
-                       roughness: float = 0.4) -> UsdShade.Material:
+                       metallic: float = 0.0,
+                       roughness: float = 1.0,
+                       identifier: Optional[str] = None) -> UsdShade.Material:
 
-    material_key = f"/material_{unique_id()}"
+    identifier = identifier or color_identifier(color, opacity)
+    material_key = f"/material_{identifier}"
     material = UsdShade.Material.Define(stage, material_key)
     shader_key = f"{material_key}/PBRShader"
     pbr_shader = UsdShade.Shader.Define(stage, shader_key)
@@ -23,3 +27,11 @@ def material_for_color(stage: Usd.Stage,
     material.CreateSurfaceOutput().ConnectToSource(pbr_shader.ConnectableAPI(), "surface")
 
     return material
+
+
+def material_for_mesh(mesh: UsdGeom.Mesh) -> UsdShade.Material:
+    prim = mesh.GetPrim()
+    relationship = prim.GetRelationship("material:binding")
+    target = relationship.GetTargets()[0]
+    material_prim = prim.GetStage().GetPrimAtPath(target)
+    return UsdShade.Material(material_prim)
